Move wireup to local class to prevent circular dependencies. Remove old duplicate test. Rename tests for better regex. E.g. `go test -test.v -test.run=".*Component.*"`
