+++ sockshop/payment/component_test.go
@@ -1,10 +1,8 @@
-package payment_component_test
+package payment
 
 import (
 	"bytes"
 	"encoding/json"
-	"github.com/weaveworks/microservices-demo/sockshop/payment"
-	"github.com/weaveworks/microservices-demo/sockshop/payment/cmd/paymentsvc"
 	"golang.org/x/net/context"
 	"io/ioutil"
 	"net/http"
@@ -16,12 +14,12 @@ func TestComponent(t *testing.T) {
 	// Mechanical stuff.
 	ctx := context.Background()
 
-	handler, logger := main.Handler(ctx, float32(99.99))
+	handler, logger := WireUp(ctx, float32(99.99))
 
 	ts := httptest.NewServer(handler)
 	defer ts.Close()
 
-	var request payment.AuthoriseRequest
+	var request AuthoriseRequest
 	request.Amount = 9.99
 	requestBytes, err := json.Marshal(request)
 	if err != nil {
@@ -37,7 +35,7 @@ func TestComponent(t *testing.T) {
 	if err != nil {
 		logger.Log("ERROR", err)
 	}
-	var response payment.Authorisation
+	var response Authorisation
 	json.Unmarshal(greeting, &response)
 
 	logger.Log("Authorised", response.Authorised)
